adjustments to comments
