+++ scripts/components/main-section.component.js
@@ -1,4 +1,4 @@
-import { ElementBuilder } from './element-builder.js';
+import { ElementBuilder } from '../utilities/element-builder.js';
 
 export class MainSectionComp {
     #arrowDown = `<svg width="43" height="60" viewBox="0 0 43 60" fill="none" xmlns="http://www.w3.org/2000/svg"><path d="M1 33L21 58M21 58L41.5 32M21 58V0" stroke-width="2" /></svg>`;
