+++ scripts/components/slider.component.js
@@ -1,5 +1,5 @@
-import { ElementBuilder } from './element-builder.js';
-import { SliderFrameComp } from './slider-frame-component.js';
+import { ElementBuilder } from '../utilities/element-builder.js';
+import { SliderFrameComp } from './slider-frame.component.js';
 
 export class SliderComp {
     #content;
