+++ scripts/components/wrapper.component.js
@@ -1,16 +1,14 @@
-import { ElementBuilder } from './element-builder.js';
-import { MainSectionComp } from './main-section-component.js';
-import { MovieSectionComp } from './movie-section.js';
-import { SignUpComp } from './signup-component.js';
-import { SliderComp } from './slider-component.js';
+import { ElementBuilder } from '../utilities/element-builder.js';
+import { MainSectionComp } from './main-section.component.js';
+import { MovieSectionComp } from './movie-section.component.js';
+import { SignUpComp } from './signup.component.js';
+import { SliderComp } from './slider.component.js';
 
 export class WrapperComp {
-    #createChildrenArr() {
-        let childrenArr = [];
+    #wrapperChildren = [
+        new MainSectionComp(300).render(), // TODO why not just define components in array literal instead of pushing them everytime?   Corrected
 
-        childrenArr.push(new MainSectionComp(300).render()); // TODO why not just define components in array literal instead of pushing them everytime?
-
-        childrenArr.push(new MovieSectionComp(
+        new MovieSectionComp(
             {
                 sectionClass: 'straight-direction-description',
                 position: '10',
@@ -19,12 +17,12 @@ export class WrapperComp {
                 imgSrc: 'images/guardians.jpg',
                 imgAlt: 'guardians of the galaxy',
                 about: `While the Awesome Mix Vol. 1 in Guardians of the Galaxy was resonant with a lot of people, it was the soundtrack in Guardians
-                of the Galaxy Vol. 2 that improved on the formula. The first film featured songs that were
-                fun and upbeat but didn't have much to do with the film's story.`
+                        of the Galaxy Vol. 2 that improved on the formula. The first film featured songs that were
+                        fun and upbeat but didn't have much to do with the film's story.`
             }
-        ).render());
+        ).render(),
 
-        childrenArr.push(new MovieSectionComp(
+        new MovieSectionComp(
             {
                 sectionClass: 'reverse-direction-description',
                 position: '09',
@@ -33,12 +31,12 @@ export class WrapperComp {
                 imgSrc: 'images/jurassic.jpg',
                 imgAlt: 'jurassic park',
                 about: `John Williams did a lot of music for many popular franchises. After his work on Star Wars, he would later do the score for
-                Jurassic Park. This dinosaur film was full of epic shots and tense moments that were further
-                brought to life by Williams' music.`
+                        Jurassic Park. This dinosaur film was full of epic shots and tense moments that were further
+                        brought to life by Williams' music.`
             }
-        ).render());
+        ).render(),
 
-        childrenArr.push(new MovieSectionComp(
+        new MovieSectionComp(
             {
                 sectionClass: 'central-direction-description',
                 imgClass: 'star-wars',
@@ -46,12 +44,12 @@ export class WrapperComp {
                 name: 'STAR WARS: A NEW HOPE',
                 audioName: 'star-wars-a-new-hope',
                 about: `When Star Wars: A New Hope was released, it introduced many iconic themes that people would recognize decades after. That
-                was thanks to John Williams, who put together the iconic fanfare, the Imperial March, and
-                so many more great tracks.`
+                        was thanks to John Williams, who put together the iconic fanfare, the Imperial March, and
+                        so many more great tracks.`
             }
-        ).render());
+        ).render(),
 
-        childrenArr.push(new SliderComp([
+        new SliderComp([
             {
                 src: 'videos/star-wars-a-new-hope.mp4',
                 imgSrc: 'images/conference_room.jpg',
@@ -69,9 +67,9 @@ export class WrapperComp {
                 imgSrc: 'images/little_tree.jpg',
                 imgAlt: 'little tree presses a button'
             }
-        ], 1).render());
+        ], 1).render(),
 
-        childrenArr.push(new MovieSectionComp(
+        new MovieSectionComp(
             {
                 sectionClass: 'straight-direction-description',
                 position: '07',
@@ -80,12 +78,12 @@ export class WrapperComp {
                 imgSrc: 'images/baby_driver.jpg',
                 imgAlt: 'baby-driver',
                 about: `Baby Driver's soundtrack is similar to Guardians of the Galaxy in many ways. It uses a lot of older songs to provide a backdrop
-                to the film's many beats. However, what Edgar Wright did with the music was so far beyond
-                that.`
+                        to the film's many beats. However, what Edgar Wright did with the music was so far beyond
+                        that.`
             }
-        ).render());
+        ).render(),
 
-        childrenArr.push(new MovieSectionComp(
+        new MovieSectionComp(
             {
                 sectionClass: 'reverse-direction-description',
                 position: '06',
@@ -94,13 +92,13 @@ export class WrapperComp {
                 imgSrc: 'images/goodfellas.jpg',
                 imgAlt: 'goodfellas',
                 about: `Martin Scorcese's movie Goodfellas remains one of his best to date. The movie deals with gangs, drugs, and everything else
-                in between. It's a crime movie that isn't afraid to deal with the dark side of life. Going
-                along with every scene is a great soundtrack full of hand-picked songs that compliment every
-                moment they appear in.`
+                        in between. It's a crime movie that isn't afraid to deal with the dark side of life. Going
+                        along with every scene is a great soundtrack full of hand-picked songs that compliment every
+                        moment they appear in.`
             }
-        ).render());
+        ).render(),
 
-        childrenArr.push(new MovieSectionComp(
+        new MovieSectionComp(
             {
                 sectionClass: 'central-direction-description',
                 imgClass: 'runner',
@@ -108,12 +106,12 @@ export class WrapperComp {
                 name: 'BLADE RUNNER',
                 audioName: 'blade-runner',
                 about: `It's astounding that Blade Runner didn't become as popular as other movies released in its time. It arguably has one of the
-                best soundtracks in movie history, with every tune being a perfect match with the action
-                on-screen.`
+                        best soundtracks in movie history, with every tune being a perfect match with the action
+                        on-screen.`
             }
-        ).render());
+        ).render(),
 
-        childrenArr.push(new SliderComp([
+        new SliderComp([
             {
                 src: 'videos/blade-runner.mp4',
                 imgSrc: 'images/bladerunner.jpg',
@@ -131,9 +129,9 @@ export class WrapperComp {
                 imgSrc: 'images/Baby-Driver_driver.jpg',
                 imgAlt: 'driver'
             }
-        ], 1).render());
+        ], 1).render(),
 
-        childrenArr.push(new MovieSectionComp(
+        new MovieSectionComp(
             {
                 sectionClass: 'straight-direction-description',
                 position: '04',
@@ -142,12 +140,12 @@ export class WrapperComp {
                 imgSrc: 'images/o-brother.jpg',
                 imgAlt: 'o brother where art thou',
                 about: `O Brother, Where Art Thou? is a movie that fires on all cylinders. It takes place in the Great Depression and involves a
-                group of convicts who go on a wild journey to find a treasure of sorts. With this film based
-                in a stylistic period in history, the soundtrack was designed to match it.`
+                        group of convicts who go on a wild journey to find a treasure of sorts. With this film based
+                        in a stylistic period in history, the soundtrack was designed to match it.`
             }
-        ).render());
+        ).render(),
 
-        childrenArr.push(new MovieSectionComp(
+        new MovieSectionComp(
             {
                 sectionClass: 'reverse-direction-description',
                 position: '03',
@@ -156,12 +154,12 @@ export class WrapperComp {
                 imgSrc: 'images/davebowman.jpg',
                 imgAlt: 'space odyssey',
                 about: `The movie tries very hard to sell the idea of what space exploration would be like, and its themes of isolation and sophistication
-                are further enhanced by its soundtrack. 2001: A Space Odyssey makes use of classical themes
-                and motifs to narrow down a tone that makes the movie feel all its own.`
+                        are further enhanced by its soundtrack. 2001: A Space Odyssey makes use of classical themes
+                        and motifs to narrow down a tone that makes the movie feel all its own.`
             }
-        ).render());
+        ).render(),
 
-        childrenArr.push(new MovieSectionComp(
+        new MovieSectionComp(
             {
                 sectionClass: 'central-direction-description',
                 imgClass: 'godfuther',
@@ -169,12 +167,12 @@ export class WrapperComp {
                 name: 'THE GODFATHER',
                 audioName: 'the-godfather',
                 about: `The Godfather is one of cinema's best works. There are so many pieces in that movie that just work, and the soundtrack is
-                part of it. Because the movie deals with crime, gangs, and the works, the music is designed
-                to reflect that.`
+                        part of it. Because the movie deals with crime, gangs, and the works, the music is designed
+                        to reflect that.`
             }
-        ).render());
+        ).render(),
 
-        childrenArr.push(new SliderComp([
+        new SliderComp([
             {
                 src: 'videos/o-brother-where-art-thou.mp4',
                 imgSrc: 'images/o-brother-image.jpg',
@@ -192,9 +190,9 @@ export class WrapperComp {
                 imgSrc: 'images/amanda.jpg',
                 imgAlt: 'amanda from a space odyssey'
             }
-        ], 1).render());
+        ], 1).render(),
 
-        childrenArr.push(new MovieSectionComp(
+        new MovieSectionComp(
             {
                 sectionClass: 'straight-direction-description',
                 position: '01',
@@ -203,18 +201,16 @@ export class WrapperComp {
                 imgSrc: 'images/frodo.jpg',
                 imgAlt: 'Frodo and the ring',
                 about: `Everything about the soundtrack in The Lord of the Rings is excellent, which is one of the many reasons that the trilogy
-                remains one of the most beloved in cinema history. Where Peter Jackson had a frame of reference
-                with Tolkien's detailed descriptions, Howard Shore had to match those visuals with music
-                all his own.`
+                        remains one of the most beloved in cinema history. Where Peter Jackson had a frame of reference
+                        with Tolkien's detailed descriptions, Howard Shore had to match those visuals with music
+                        all his own.`
             }
-        ).render());
+        ).render(),
 
-        childrenArr.push(new SignUpComp().render());
-
-        return childrenArr;
-    }
+        new SignUpComp().render()
+    ];
 
     render() {
-        return new ElementBuilder('main').setChildren(...this.#createChildrenArr()).build();
+        return new ElementBuilder('main').setChildren(...this.#wrapperChildren).build();
     }
 }
