+++ scripts/services/animation.service.js
