+++ scripts/utilities/element-builder.js
@@ -1,4 +1,4 @@
-export class ElementBuilder { // TODO: this is not a component !!! this is utility!!!
+export class ElementBuilder { // TODO: this is not a component !!! this is utility!!!    Corrected
     #tagName;
     #classes;
     #attributes;
